+++ docs/src/app/app-root.tsx
@@ -71,12 +71,13 @@ export class AppRoot {
 
           <ion-route url="/deck" component="app-deck-transition" />
           <ion-route url="/deck/transition" component="app-deck-transition" />
+          <ion-route url="/deck/background" component="app-deck-background" />
+          <ion-route url="/deck/header-footer" component="app-deck-header-footer" />
           <ion-route url="/deck/navigation" component="app-deck-navigation" />
           <ion-route url="/deck/pager" component="app-deck-pager" />
           <ion-route url="/deck/size" component="app-deck-size" />
           <ion-route url="/deck/extra" component="app-deck-extra-features" />
           <ion-route url="/deck/events" component="app-deck-events" />
-          <ion-route url="/deck/background" component="app-deck-background" />
           <ion-route url="/deck/actions" component="app-deck-actions" />
 
           <ion-route url="/misc" component="app-misc-open-source" />
@@ -239,6 +240,9 @@ export class AppRoot {
                   </ion-item>
                   <ion-item detail={false} href="/deck/background" routerDirection="forward">
                     <ion-label>Background</ion-label>
+                  </ion-item>
+                  <ion-item detail={false} href="/deck/header-footer" routerDirection="forward">
+                    <ion-label>Header &amp; Footer</ion-label>
                   </ion-item>
                   <ion-item detail={false} href="/deck/navigation" routerDirection="forward">
                     <ion-label>Navigation</ion-label>